+++ modules/aws/eks-global/main.tf
@@ -12,8 +12,6 @@ terraform {
   }
 }
 
-data "aws_caller_identity" "current" {}
-
 data "aws_region" "current" {}
 
 locals {